refactor: moved type definition to same package where it is used
