+++ cqp-api/src/main/java/com/quantori/cqp/api/service/LibraryAlreadyExistsException.java
@@ -1,4 +1,4 @@
-package com.quantori.cqp.api.model;
+package com.quantori.cqp.api.service;
 
 /**
  * An error message to indicate that a requested to be created library already exists. The API does not add any
